+++ src/modules/diginext/dx-api.ts
@@ -28,6 +28,7 @@ export async function dxApi<T = ResponseData>(options: AxiosRequestConfig & { dx
 		return responseData as T;
 	} catch (e) {
 		console.log("e :>> ", e);
-		return { status: 0, messages: [`${e}`] } as T;
+		const err: string = e.response || e.data?.message === "UNAUTHORIZED" || e.data?.status === 401 ? "Invalid DX Key." : e.message;
+		return { status: 0, messages: [`${err}`] } as T;
 	}
 }