+++ src/routes/auth/google/index.ts
@@ -12,7 +12,7 @@ const router = express.Router();
 
 // http://localhost:6969/auth/google?redirect_url=http://localhost:6969/auth/profile
 
-const signAndRedirect = (res: Response, data: { userId: string; workspaceId?: string }, redirectUrl: string) => {
+export const signAndRedirect = (res: Response, data: { userId: string; workspaceId?: string }, redirectUrl: string) => {
 	const { userId, workspaceId } = data;
 	// console.log("[2] signAndRedirect > data :>> ", data);
 	const access_token = generateJWT(userId, { expiresIn: process.env.JWT_EXPIRE_TIME || "2d", workspaceId });